requirements.txt
fastapi
pydantic==1.10.11
pydantic_core==2.3.0
SQLAlchemy
alembic
uvicorn
psycopg2
python-dotenv